WIP! Refactor BAMSingleEnd into BAMSubset
Experimenting with how this might work as a way to have arbitrary
subsets.  One complication that arises is handling subsets (like mapped
reads) where the count is more efficiently known by subtraction of
unmatching reads.  Another, larger issues, is that subsets start to have
multiple parents.  In the case of single-end mapped reads, for example,
there's both the BAMSingleEnd from the same dataset and the BAMSubset
(mapped) of the parent dataset.

This may be the way to go, but for now it's just additional
complication.  (Does HDF5 help with this at all?)