Switch to a more traditional option syntax of --stage instead of :::
::: is a bit strange and unusual, and there's really no need for it.

I'd have liked to use Getopt::Long inside parse_args() to handle
--stage, but I couldn't figure out how to make it push each occurrence
of the option into a separate array ref.  The closest I got was:

    'stage=s{2,4}' => \@stages,

but @stages is a flat array of all stage names and files.  A callback is
called once per-value instead of once per-option, and there's no good
way to discern which value you're looking at.